Unify pruning var implementations
