+++ community/cypher/cypher-planner-3.4/src/main/scala/org/neo4j/cypher/internal/compiler/v3_4/planner/logical/plans/rewriter/pruningVarExpander.scala
@@ -100,11 +100,7 @@ case object pruningVarExpander extends Rewriter {
 
         val innerRewriter = topDown(Rewriter.lift {
           case expand@VarExpand(lhs, fromId, dir, _, relTypes, toId, _, length, ExpandAll, _, _, _, _, predicates) if distinctSet(expand.selfThis) =>
-            if (length.min >= 4 && length.max.get >= 5)
-              // These constants were selected by benchmarking on randomized graphs, with different
-              // degrees of interconnection.
-              FullPruningVarExpand(lhs, fromId, dir, relTypes, toId, length.min, length.max.get, predicates)(expand.solved)(SameId(expand.id))
-            else if (length.max.get > 1)
+            if (length.max.get > 1)
               PruningVarExpand(lhs, fromId, dir, relTypes, toId, length.min, length.max.get, predicates)(expand.solved)(SameId(expand.id))
             else expand.selfThis
         })
