+++ community/values/src/main/java/org/neo4j/values/virtual/RelationshipValue.java
@@ -59,7 +59,7 @@ public abstract class RelationshipValue extends VirtualRelationshipValue
 
     public abstract MapValue properties();
 
-    public NodeValue otherNode( NodeValue node )
+    public NodeValue otherNode( VirtualNodeValue node )
     {
         return node.equals( startNode() ) ? endNode() : startNode();
     }
