+++ enterprise/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/CypherComparisonSupport.scala
@@ -346,8 +346,8 @@ object CypherComparisonSupport {
 
     implicit def versionToVersions(version: Version): Versions = Versions(version)
 
-    val oldest = orderedVersions.head
-    val latest = orderedVersions.last
+    val oldest: Version = orderedVersions.head
+    val latest: Version = orderedVersions.last
     val all = Versions(orderedVersions: _*)
 
     object V2_3 extends Version("2.3")
